server/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
openai>=1.0.0
httpx>=0.25.0
Pillow>=10.0.0
